refactor(websocket): move useInitializeWebsocket hook into its own file
