+++ frontend/hooks/useChat.ts
@@ -1,11 +1,5 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { useChatStore } from '@/store/useChatStore';
-import { useSidebarStore } from '@/store/useSidebarStore';
-import { useAuthStore } from '@/store/useAuthStore';
-
-import { toast } from 'sonner';
-
-import { themes } from '@/lib/themes';
 
 import {
   sendMessage,
@@ -16,9 +10,6 @@ import {
   getConversationTheme,
   markConversationAsRead,
 } from '@/lib/api/messages';
-import { useNotificationsStore } from '@/store/useNotificationsStore';
-
-const fastapiWebsocketUrl = process.env.NEXT_PUBLIC_FASTAPI_WEBSOCKET_URL;
 
 export function useSendMessage() {
   const [loading, setLoading] = useState(false);
@@ -275,168 +266,3 @@ export function useMarkConversationAsRead() {
 
   return { markAsRead, loading, error };
 }
-
-export const useInitializeWebsocket = () => {
-  const clearChat = useChatStore((state) => state.clearChat);
-
-  const addMessage = useChatStore((state) => state.addMessage);
-
-  const setSocket = useChatStore((state) => state.setSocket);
-
-  const setOtherParticipantFriendshipStatus = useChatStore(
-    (state) => state.setOtherParticipantFriendshipStatus
-  );
-
-  const setConversationTheme = useChatStore((state) => state.setConversationTheme);
-
-  const setConversationThemeChangedBy = useChatStore(
-    (state) => state.setConversationThemeChangedBy
-  );
-
-  const setConversationThemeChangedAt = useChatStore(
-    (state) => state.setConversationThemeChangedAt
-  );
-
-  const setOtherParticipantIsOnline = useChatStore((state) => state.setOtherParticipantIsOnline);
-
-  const setOtherParticipantLastOnline = useChatStore(
-    (state) => state.setOtherParticipantLastOnline
-  );
-
-  const setOtherParticipantLastReadMessageId = useChatStore(
-    (state) => state.setOtherParticipantLastReadMessageId
-  );
-
-  const setOtherParticipantLastReadMessageAt = useChatStore(
-    (state) => state.setOtherParticipantLastReadMessageAt
-  );
-
-  const removeSendingMessage = useChatStore((state) => state.removeSendingMessage);
-
-  const currentUserId = useAuthStore((state) => state.currentUserId);
-
-  const upsertSnippet = useSidebarStore((state) => state.upsertSnippet);
-
-  const updateHasSeenLatestMessage = useSidebarStore((state) => state.updateHasSeenLatestMessage);
-
-  const getNotifications = useNotificationsStore((state) => state.getNotifications);
-
-  const { markAsRead } = useMarkConversationAsRead();
-
-  useEffect(() => {
-    if (!currentUserId) return;
-
-    // Start new connection
-    const ws = new WebSocket(`${fastapiWebsocketUrl}/api/websocket/ws`);
-
-    ws.onopen = () => {
-      console.log('Connected to:', currentUserId);
-    };
-
-    ws.onmessage = (event) => {
-      const eventData = JSON.parse(event.data);
-
-      // getState() is used because onmessage is a callback function
-      const latestActiveConversationId = useChatStore.getState().activeConversationId;
-
-      if (eventData.type === 'NEW_MESSAGE') {
-        if (eventData.data.conversationId === latestActiveConversationId) {
-          addMessage(eventData.data);
-
-          if (eventData.data.senderId === currentUserId) {
-            removeSendingMessage(eventData.data.tempMessageId);
-          } else if (latestActiveConversationId) {
-            markAsRead(latestActiveConversationId);
-          }
-        }
-        upsertSnippet(eventData.data);
-      } else if (eventData.type === 'UPDATE_CONVERSATION') {
-        setOtherParticipantFriendshipStatus(eventData.data.friendshipStatus);
-
-        const isViewingConversations = useChatStore.getState().activeConversationId != null;
-
-        if (eventData.data.friendshipStatus === 'accepted') {
-          toast.success(
-            `You are now friends with ${eventData.data.otherParticipantUsername} again. ${isViewingConversations ? 'You can now send messages to each other.' : ''}`
-          );
-        } else {
-          toast.info(
-            `${eventData.data.otherParticipantUsername} has unfriended you. ${isViewingConversations ? 'Your conversation with them is set to read-only.' : ''}`
-          );
-        }
-      } else if (eventData.type == 'NEW_THEME') {
-        setConversationTheme(eventData.data.theme);
-        setConversationThemeChangedBy(eventData.data.changedBy);
-        setConversationThemeChangedAt(new Date(eventData.data.changedAt));
-
-        const selectedTheme = themes.find((theme) => theme.id == eventData.data.theme);
-
-        toast.success(
-          `The theme for this conversation has recently been set to ${selectedTheme?.label} by ${eventData.data.changedBy}.`
-        );
-      } else if (eventData.type === 'USER_ONLINE') {
-        setOtherParticipantIsOnline(eventData.data.isOnline);
-      } else if (eventData.type === 'USER_OFFLINE') {
-        setOtherParticipantIsOnline(eventData.data.isOnline);
-
-        if (eventData.data.lastOnline) {
-          const date = new Date(eventData.data.lastOnline);
-          setOtherParticipantLastOnline(date);
-        } else {
-          setOtherParticipantLastOnline(null);
-        }
-      } else if (eventData.type === 'MESSAGE_SEEN') {
-        updateHasSeenLatestMessage(
-          eventData.data.conversationId,
-          eventData.data.hasSeenLatestMessage,
-          eventData.data.latestMessageSenderId
-        );
-
-        setOtherParticipantLastReadMessageId(eventData.data.lastReadMessageId);
-
-        if (eventData.data.lastReadMessageAt) {
-          const date = new Date(eventData.data.lastReadMessageAt);
-          setOtherParticipantLastReadMessageAt(date);
-        } else {
-          setOtherParticipantLastReadMessageAt(null);
-        }
-      } else if (eventData.type === 'NEW_NOTIFICATION') {
-        getNotifications(false, true);
-      }
-    };
-
-    clearChat();
-
-    setSocket(ws);
-
-    const interval = setInterval(() => {
-      if (ws.readyState === WebSocket.OPEN) {
-        ws.send(JSON.stringify({ type: 'PING' }));
-      }
-    }, 30000);
-
-    return () => {
-      if (ws.readyState === WebSocket.OPEN || ws.readyState === WebSocket.CONNECTING) {
-        console.log('Cleaning up connection for:', currentUserId);
-        ws.close();
-      }
-
-      clearInterval(interval);
-    };
-  }, [
-    currentUserId,
-    setSocket,
-    addMessage,
-    clearChat,
-    upsertSnippet,
-    setOtherParticipantFriendshipStatus,
-    setConversationTheme,
-    setConversationThemeChangedAt,
-    setConversationThemeChangedBy,
-    setOtherParticipantIsOnline,
-    setOtherParticipantLastOnline,
-    updateHasSeenLatestMessage,
-    setOtherParticipantLastReadMessageAt,
-    setOtherParticipantLastReadMessageId,
-  ]);
-};
